Refactor: moved signature to signaturescheme package
